companies.id when changed from mediumint to tinyint fk fails

+++ sql/pp.sql
@@ -101,7 +101,7 @@ commit;
 -- companies
 create table companies (
     tid smallint unsigned not null,
-    id mediumint unsigned not null auto_increment,
+    id tinyint unsigned not null auto_increment,
 		primary key (id, tid), 
 
     longname varchar(50) not null,
@@ -122,7 +122,7 @@ create table companies (
 
 create table company_addresses (
     tid smallint unsigned not null,
-    company_id mediumint unsigned not null,
+    company_id tinyint unsigned not null,
 		id tinyint unsigned not null auto_increment,
 
     address varchar(200),
@@ -137,7 +137,7 @@ create table company_addresses (
 
 create table company_ibans (
     tid smallint unsigned not null,
-    company_id mediumint unsigned not null,
+    company_id tinyint unsigned not null,
 
     iban char(34), -- International Bank Account Number
     bankname varchar(50),
